# Improve String#first and #last performance