Removes unnecessary conditional and method call for significant performance improvement.  As a side effect, this fixes an unexpected behavior where passing a limit of 0 would return a frozen string.

This also implements the Rails 6.1 intended behavior with regards to negative limits, and removes the previous deprecation warnings.

```
String#first Comparison:
  new:   3056515.0 i/s
  old:   1943310.2 i/s - 1.57x slower

String#last Comparison:
  new:   2691919.0 i/s
  old:   1924256.6 i/s - 1.40x slower
```

(Note: "old" benchmarks have deprecation warnings commented out, for a more fair comparison.)