@schneems 2019-05-05
Is master tracking 6.0 or 6.1? We cannot remove that deprecation until 6.1.