@kaspth 2019-05-05
The future release manager usually removes deprecations to keep potential backports easier. I’ll defer that decision to @rafaelfranca