@jonathanhefner 2019-05-05
@schneems Sure!  You may need to `gem install benchmark-inputs` first.

```ruby
require "benchmark/inputs"

class String
  def to(position)
    self[0..position]
  end

  def from(position)
    self[position..-1]
  end

  def first_old(limit = 1)
    if limit == 0
      ""
    elsif limit >= size
      dup
    else
      to(limit - 1)
    end
  end

  def first_new(limit = 1)
    self[0, limit] || (raise ArgumentError, "negative limit")
  end

  def last_old(limit = 1)
    if limit == 0
      ""
    elsif limit >= size
      dup
    else
      from(-limit)
    end
  end

  def last_new(limit = 1)
    self[[length - limit, 0].max, limit] || (raise ArgumentError, "negative limit")
  end
end


STRING = "abcdef"

Benchmark.inputs((0..STRING.length + 1).to_a) do |job|
  job.report("String#first old"){|limit| STRING.first_old(limit) }
  job.report("String#first new"){|limit| STRING.first_new(limit) }
  job.compare!
end

Benchmark.inputs((0..STRING.length + 1).to_a) do |job|
  job.report("String#last old"){|limit| STRING.last_old(limit) }
  job.report("String#last new"){|limit| STRING.last_new(limit) }
  job.compare!
end
```