@jonathanhefner 2019-05-05
@kaspth Understood.  Without raising the error, the optimized version would simply return `nil`.  But I could add a conditional like so if you think it's more appropriate:

```ruby
def first(limit = 1)
  if limit < 0
    ActiveSupport::Deprecation.warn(
      "Calling String#first with a negative integer limit " \
      "will raise an ArgumentError in Rails 6.1."
    ) 
    to(limit - 1)
  else
    self[0, limit]
  end
end
```